Refactor: Move laundrocleanservices types into a folder
